Move glossary to data folder
