+++ BQE_api/BQE_api/urls.py
@@ -3,5 +3,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path('bible/', include('bible.urls')),
+    path('glossary/', include('glossary.urls')),
     path('admin/', admin.site.urls),
 ]
