+++ BQE_api/glossary/data/book_abbreviations.py
