+++ BQE_api/glossary/urls.py
@@ -1,8 +1,7 @@
 from django.urls import path
-
 from .views import *
 
 urlpatterns = [
-    path('', index, name='index')
-
+    path('', index, name='index'),
+    path('seed_glossary', seed_glossary, name='seed_glossary'),
 ]
